move fs to internal/fs
